1.0.37 beta

+++ app/Http/Controllers/DepartmentController.php
@@ -9,7 +9,7 @@ class DepartmentController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return array|\Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
      */
     public function index()
     {
@@ -28,7 +28,7 @@ class DepartmentController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Http\RedirectResponse|\Illuminate\Http\Response|\Illuminate\Routing\Redirector
+     * @return array|\Illuminate\Contracts\Foundation\Application|\Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
      */
     public function create(Request $request)
     {
@@ -76,10 +76,12 @@ class DepartmentController extends Controller
      * Show the form for editing the specified resource.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return array|\Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
      */
     public function edit($id)
     {
+        abort_unless(auth()->user()->can('edit_department'),403,'you dont have required authorization to this resource');
+
         try {
             $department=Department::find($id);
             return view('department.editform',compact('department'));
@@ -97,10 +99,12 @@ class DepartmentController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return array|\Illuminate\Contracts\Foundation\Application|\Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
      */
     public function update(Request $request, $id)
    {
+       abort_unless(auth()->user()->can('edit_department'),403,'you dont have required authorization to this resource');
+
        try {
            $department=Department::find($id);
            $department->dept_name=$request->dept_name;
@@ -141,10 +145,12 @@ class DepartmentController extends Controller
      * Remove the specified resource from storage.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return array|\Illuminate\Contracts\Foundation\Application|\Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
      */
     public function destroy($id)
     {
+        abort_unless(auth()->user()->can('delete_department'),403,'you dont have required authorization to this resource');
+
         try {
             Department::destroy($id);
             return redirect(route('index'))->with('success','Department Delete Successfully');
